+++ Client.Test/Internal/AssemblyHelperTest.cs
@@ -12,7 +12,7 @@ namespace InfluxDB3.Client.Test.Internal
             var version = AssemblyHelper.GetVersion();
             Assert.Multiple(() =>
             {
-                Assert.That(Version.Parse(version).Major, Is.EqualTo(0));
+                Assert.That(Version.Parse(version).Major, Is.EqualTo(1));
                 Assert.That(Version.Parse(version).Minor, Is.GreaterThanOrEqualTo(0));
                 Assert.That(Version.Parse(version).Build, Is.EqualTo(0));
                 Assert.That(Version.Parse(version).Revision, Is.EqualTo(0));