+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.0 [unreleased]
+
+### Features
+
+1. [#132](https://github.com/InfluxCommunity/influxdb3-csharp/pull/132): Respect iox::column_type::field metadata when
+   mapping query results into values.
+    - iox::column_type::field::integer: => Long
+    - iox::column_type::field::uinteger: => Long
+    - iox::column_type::field::float: => Double
+    - iox::column_type::field::string: => String
+    - iox::column_type::field::boolean: => Boolean
+
 ## 0.9.0 [unreleased]
 
 ## 0.8.0 [2024-09-13]